beniCulturali_Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
click
Faker